generator: Refactor ModelResolver to use IModelModelMapper
This was an existing point of abstraction created earlier and expressly
for this purpose, but I guess I just forgot to perform the refactor
earlier. This interface facilitates the future creation of new
transformations on the model graph that map one existing model
transparently to another, and can be used, e.g., to "erase" models from
the graph.