generator: Use repository for reverse model resolution
Previously, model-type mappers were required to support a reverse
operation--a query that would produce the schema model name given the
Rust type. I had been putting off this refactor for a while, but now
that there's a need to map inline schema names to types (by definition a
lossy function), this constraint is no longer possible to enforce for
all model-type mappers. Instead, implement a repository "cache" in the
ModelResolver to allow these two to be associated the first time they're
resolved in the forward direction.