generator: Remove DocumentParser
This code, which was borrowed from swagger-codegen, flattens inline
models to become first-class schemas. It was very useful when there were
a number of inline schemas that had properties and required additional
struct definitions. In 2024.2, however, there are many inline structs
that can be rendered simply using std::option::Option, and so this code
has become more of a hindrance.