generator: Promoted inline schemas
This logic replaces the old DocumentParser class. Inline schemas that
represent algebraic product types, e.g. structs with members, need to be
promoted to "first-class" schemas and given a definition and schema ref,
so that their definitions can be generated. This commit introduces the
logic to do this in OpenapiSpecification, and a model-type mapper that
resolves the created $refs to types.