generator: Resolve some inline schemas to std types
There are a number of inline schemas in the Redfish data model that
represent simple types--for example, there's a repeated pattern that
specifies a property may either be of some type (where a $ref is
provided) or "null". Obviously, we already have a concept for that
(Option<>), so create a resolver that can recognize these primitive
patterns.