generator: Don't abort if there's no JSON schema file
Previously, we would throw a RuntimeException if we couldn't identify
the JSON Schema file for a particular model. This is because in the
general case, services must indicate the JSON schema file for a model in
the headers of a response. However, some models (e.g.
odata_v4::ServiceDocument) actually aren't represented in the JSON
schema. Now, we allow these models to be generated without
implementations for Metadata<'static>.