Have generate.mk check explicitly for GNU mv
generate.mk depends on the '--update=none' flag that is only provided by recent
versions of coreutils. For operating systems that do not provide a GNU mv at
all, or those that only have older versions available, an updated coreutils can
be installed by homebrew, which applies a 'g' prefix to the command that we can
test for.